Streamlined 4 functions into 1
instead of summing each corner, just have the locations passed into one function which sums the values in those locations.  removed as much of the appending to a list and made list comprehension or just passed straight into another function as an iterable.

Added a name==main to make the script more like a proper program.  changed global variables to ALL CAPS as this is a standard that helps other devs read your code.. their is global variables and local, local is specific to each function while global variables are accessed from anywhere in your script.